buttons of doom
A now-notorious fanfiction created by Peter Chimaera, noted for being a
representation of the human condition fucking hilarious. It is so easily
adaptable and so ubiquitous that it's become one of the most beloved
Memes on the Internet, achieving a popularity, even on /tg/, that's
seldom seen from memes that arrive on it as migrants. Taglines from it
are eminently quotable and led to literally thousands of variations of
the story, which means, that it's a rare case of pretty much everyone
indulging in something for the sake of its silliness.